requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.3
djangorestframework==3.14.0
drf-spectacular==0.24.2
et-xmlfile==1.1.0
inflection==0.5.1
Jinja2==3.1.2
jsonschema==4.17.1
MarkupSafe==2.1.1
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
